Added home#index routed root there to dummy application, and adjusted layout accordingly.
I forgot earlier to use main_app. and restbook. before URL helpers in the layout.
All URL helpers in a layout shared between the engine and host app should
only use URL helpers namespaced in this way (or else, for example, a layout's
link meant to go to the main app's root would instead go to the engine's root
when rendering engine views).